+++ CHANGES.rst
@@ -227,6 +227,8 @@ astropy.wcs
   (https://doi.org/10.5281/zenodo.1188875) as well as the high-level wrapper
   class for low-level WCS objects. [#7326]
 
+- Added a new property ``WCS.has_distortion``. [#7326]
+
 API Changes
 -----------